+++ src/hooks/useSwap.tsx
@@ -29,7 +29,11 @@ import orderInProgressStore from "../store/orderInProgressStore";
 import pendingOrdersStore from "../store/pendingOrdersStore";
 import BigNumber from "bignumber.js";
 import { useSolanaWallet } from "./useSolanaWallet";
-import { formatAmount, getOrderPair } from "../utils/utils";
+import {
+  formatAmount,
+  formatAmountInNumber,
+  getOrderPair,
+} from "../utils/utils";
 import { useNetworkFees } from "./useNetworkFees";
 import { useSuiWallet } from "./useSuiWallet";
 import logger from "../utils/logger";
@@ -99,7 +103,7 @@ export const useSwap = () => {
             inputAsset.decimals,
             Math.min(inputAsset.decimals, BTC.decimals)
           )
-        : Number(inputBalance)),
+        : inputBalance.toString()),
     [inputBalance, inputAsset]
   );
 
@@ -169,12 +173,12 @@ export const useSwap = () => {
     if (!limits) return defaultLimits;
     else
       return {
-        minAmount: formatAmount(
+        minAmount: formatAmountInNumber(
           limits.minAmount,
           inputAsset.decimals,
           inputAsset.decimals
         ),
-        maxAmount: formatAmount(
+        maxAmount: formatAmountInNumber(
           limits.maxAmount,
           inputAsset.decimals,
           inputAsset.decimals
@@ -351,7 +355,7 @@ export const useSwap = () => {
         return;
       }
 
-      setError({ inputError: Errors.none });
+      setError({ inputError: Errors.none, outputError: Errors.none });
 
       if (!inputAsset || !outputAsset || !Number(amount)) return;
 
@@ -384,7 +388,7 @@ export const useSwap = () => {
       return;
     }
 
-    setError({ outputError: Errors.none });
+    setError({ inputError: Errors.none, outputError: Errors.none });
 
     if (!inputAsset || !outputAsset || !amountInNumber) return;