+++ src/utils/utils.spec.ts
@@ -1,10 +1,10 @@
 import { describe, it } from "vitest";
-import { formatAmount } from "./utils";
+import { formatAmountInNumber } from "./utils";
 import logger from "./logger";
 
 describe("utils", () => {
   it("formatNumber", () => {
-    const number = formatAmount("5000", 8);
+    const number = formatAmountInNumber("5000", 8);
     logger.log("number :", number);
   });
 });